Fixed some compiler warnings

+++ Code/Entity.cpp
@@ -87,7 +87,7 @@ namespace Monocle
 
 	int Entity::GetNumberOfTags()
 	{
-		return tags.size();
+		return static_cast<int>(tags.size());
 	}
 
 	int Entity::GetLayer()
